# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MathAgent 研究状态服务",
    "description": "本地研究工作台与受控运行接口。support 表示证据策略下的支持，不是数学真值。Fake 运行均是确定性测试。",
    "version": "0.2.0"
  },
  "paths": {
...
    "/exports": {
      "post": {
        "tags": [
          "artifacts"
        ],
        "summary": "Create Export",
        "operationId": "create_export_exports_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exports/{export_id}/download": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "Download",
        "operationId": "download_exports__export_id__download_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/examples/hermitian": {
      "post": {
        "tags": [
          "artifacts"
        ],
        "summary": "Hermitian",
        "operationId": "hermitian_examples_hermitian_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Command": {
        "properties": {},
        "additionalProperties": false,
        "type": "object",
        "title": "Command"
      },
...
      "ExportCreate": {
        "properties": {
          "project_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Project Id"
          },
          "branch_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Branch Id"
          },
          "object_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "maxLength": 100,
                  "minLength": 1
                },
                "type": "array",
                "maxItems": 10000,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "branch_id"
        ],
        "title": "ExportCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}